Support building with unmodified Android NDK

+++ src/modules/roc_packet/target_posix/roc_packet/address.cpp
@@ -27,7 +27,7 @@ bool Address::set_saddr(const sockaddr* sa) {
         return false;
     }
 
-    memcpy(&sa_, sa, sa_size);
+    memcpy(&sa_, sa, (size_t)sa_size);
 
     return true;
 }
